requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-multipart
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator